requirement.txt
setuptools
jinja2
flask
flask-script
mongoengine
blinker
mongomock
tox
flask_mongoengine
paramiko
plumbum
celery
fabric
shell
python-nmap
scikit-learn
